UI Position added

+++ src/graphics/font.rs
@@ -9,18 +9,18 @@ use super::{
 
 #[derive(Debug)]
 pub struct Font {
-    pub atlas_size: Vec2D<u32>,
-    pub chars_on_page: u32,
-    pub char_size: Vec2D<u32>,
+    pub atlas_size: Vec2D<i32>,
+    pub chars_on_page: i32,
+    pub char_size: Vec2D<i32>,
     pub st_char: Vec2D<f32>,
-    pub indent: u32,
+    pub indent: i32,
     pub pages: Pages<Texture>,
 }
 
 impl Font {
-    pub fn new<S>(atlas_size: S, indent: u32, pages: Pages<Texture>) -> Self
+    pub fn new<S>(atlas_size: S, indent: i32, pages: Pages<Texture>) -> Self
         where
-            S: Into<Vec2D<u32>>,
+            S: Into<Vec2D<i32>>,
     {
         let atlas_size = atlas_size.into();
 
@@ -38,31 +38,31 @@ impl Font {
         }
     }
 
-    pub fn text_size(&self, text: &str) -> Vec2D<u32> {
+    pub fn text_size(&self, text: &str) -> Vec2D<i32> {
         match text.chars().count() {
             0 => Vec2D::new(0, self.char_size.y),
             1 => self.char_size,
             l => Vec2D::new(
-                self.char_size.x + (l as u32 - 1) * (self.indent as u32 + self.char_size.x),
+                self.char_size.x + (l as i32 - 1) * (self.indent + self.char_size.x),
                 self.char_size.y,
             ),
         }
     }
 
-    pub fn char_rect(&self, char_num: u32, pos: Vec2D<f32>) -> Rect<f32> {
+    pub fn char_rect(&self, char_num: i32, pos: Vec2D<f32>) -> Rect<f32> {
         let x_step = (char_num * (self.char_size.x + self.indent)) as f32;
         Rect::new((pos.x + x_step, pos.y), self.char_size.cast::<f32>())
     }
 
     pub fn st_rect(&self, code: u32) -> Rect<f32> {
-        let code_at_page = code % self.chars_on_page;
+        let code_at_page = code as i32 % self.chars_on_page;
         let s = (code_at_page % self.atlas_size.x) as f32 * self.st_char.x;
         let t = (code_at_page / self.atlas_size.x) as f32 * self.st_char.y;
         Rect::new((s, t), (self.st_char.x, self.st_char.y))
     }
 
     pub fn page(&self, code: u32) -> Option<&Texture> {
-        let page_code = code / self.chars_on_page;
+        let page_code = code as i32 / self.chars_on_page;
         self.pages.get(page_code as usize)
     }
 }
